Refactor header
