+++ less/extension.less
@@ -1,4 +1,8 @@
 .ExtensionsPage.Bazaar {
+    .ExtensionsPage-header {
+        background: #5f4bb6;
+        .box-shadow(0 1px 5px @shadow-color);
+    }
     .Button {
         &.Requirement-MemoryLimit, &.Requirement-FilePermissions {
             .Button--color(@control-danger-color, @control-danger-bg);
